Add setup.cfg: skip CI build on release.
At least works on Travis, see http://docs.travis-ci.com/user/how-to-skip-a-build/
And hopefully on Jenkins when it is configured like that.

This setup.cfg setting will get picked up by a near-future zest.releaser version,
which is in the works, and which I (Maurits) am already using in development mode.