Remove folder children optimization that breaks common usage:
There are too many things that can depend on catalog indexes being accurate
while creating the new folder for this optimization to be the default.  For
example, if any indexed value requires using the catalog to filter or
aggregate some value from children, then calling `getObject()` on those
catalog brains will raise an exception.  There are many potential use cases
that would run into this, but in my case it was an automatic workflow
transition that caused reindexing workflow variables.